--- CSentiment-API/customized_stopwords.txt
@@ -1,33 +1,0 @@
-ra
-sa
-mo
-nga
-to
-ug
-og
-koy
-ma
-ka
-na
-nga
-ko
-miss
-sir
-teacher
-maestraha
-teachers
-teaches
-maestra
-you
-your
-i
-do
-really
-always
-because
-us
-ang
-mutudlo
-students
-niya
-student
