+++ CSentiment-API/stopwords.txt
@@ -1,4 +1,5 @@
 a
+all
 also
 an
 and
@@ -19,6 +20,7 @@ by
 c
 came
 can
+cannot
 case
 cases
 come
@@ -37,12 +39,14 @@ end
 ended
 ending
 ends
+even
 ever
 every
 f
 for
 four
 from
+full
 further
 furthered
 furthering
@@ -61,6 +65,7 @@ him
 himself
 his
 how
+however
 i
 if
 in
@@ -69,6 +74,7 @@ it
 its
 itself
 j
+just
 k
 l
 last
@@ -78,6 +84,9 @@ least
 less
 let
 lets
+like
+likely
+long
 m
 man
 many
@@ -85,6 +94,8 @@ may
 me
 men
 might
+more
+most
 mr
 mrs
 much
@@ -165,6 +176,9 @@ shows
 side
 sides
 since
+small
+smaller
+smallest
 so
 some
 somebody
@@ -192,6 +206,8 @@ these
 they
 thing
 things
+think
+thinks
 this
 those
 though
@@ -200,6 +216,7 @@ through
 thus
 to
 today
+together
 too
 took
 toward
@@ -271,10 +288,3 @@ ka
 na
 nga
 ko
-miss
-sir
-teacher
-maestraha
-teachers
-teaches
-maestra
