+++ frontend/src/Comment/comment.css
@@ -1,19 +1,16 @@
-.commentBox, .resultBox{
+.commentBox,.resultBox{
 padding: 20px;
 width: 500px;
 }
 .inputComment{
     padding: 5px;
     width: 80%;
-    font-size: 20px;
 }
 .btnSubmitFragment{
     margin: 5px;
 }
 #btnSubmit{
     width: 200px;
-    height: 50px;
-    
 }
 .main_div{
     margin: auto; 
@@ -22,16 +19,4 @@ width: 500px;
     border-radius: 5px;
     max-width: 50%;
     margin-top: 50px;
-}
-
-.questionaireTable tr,th{
-    width: 100%;
-    background-color: aquamarine;
-    text-align: left;
-}
-
-.questionaireDiv
-{
-    display: block;
-    border: 1px black solid;
 }
