+++ frontend/src/Utils/ApiService.js
@@ -2,7 +2,6 @@ import axios from 'axios';
 
 const POST_COMMENT_API = 'http://127.0.0.1:5000/sentimentAnalysis';
 const POST_SENTIMENT_API = 'http://127.0.0.1:5000/getSentiment';
-const GET_SECTION1_QUESTIONS_API = 'http://127.0.0.1:5000/getQuestions/';
 
 class ApiService {
     postComment(Comment){
@@ -10,10 +9,6 @@ class ApiService {
     }
     getSentiment(Comment){
         return axios.post(POST_SENTIMENT_API,Comment);
-    }
-    getQuestions(section)
-    {
-        return axios.get(GET_SECTION1_QUESTIONS_API + section);
     }
 }
