+++ frontend/src/index.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import Comment from './Evaluation Form/evaluationForm.js';
+import Test from './Test/Test.js'
+import Comment from './Comment/comment.js';
 import reportWebVitals from './reportWebVitals';
 
 ReactDOM.render(
